refactor(test): parametrize tests over log_path
